+++ src/pages/client/Auth/components/form/RegisterForm.tsx
@@ -1,57 +1,18 @@
 // @ts-nocheck
 import React from 'react';
 import Input from '../inputs/Input';
-import { FieldValues, SubmitHandler, useForm } from 'react-hook-form';
-import { yupResolver } from '@hookform/resolvers/yup';
-import * as yup from 'yup';
-import Button from '../../../../../app/components/button/Button';
-import { Axios } from '@/app/utils/api';
-import { ENDPOINTS } from '@/app/constants/endpoint';
-import { Toaster, toast as t } from 'sonner';
 
-const validationSchema = yup.object().shape({
-	fullName: yup.string().required(),
-	email: yup.string().email().required(),
-	password: yup.string().required().min(6),
-	confirmPassword: yup.string().oneOf([yup.ref('password')]),
-});
+import Button from '../../../../../app/components/button/Button';
+import useRegisterForm from './hooks/useRegisterForm';
+import { Toaster } from 'sonner';
 
 const RegisterForm = () => {
-	const {
-		register,
-		handleSubmit,
-		reset,
-		formState: { errors, isSubmitted },
-	} = useForm({
-		resolver: yupResolver(validationSchema),
-	});
-
-	function onSubmit(data) {
-		delete data.confirmPassword;
-		const response = Axios.axiosPost(ENDPOINTS.AUTH.REGISTER, data);
-
-		t.promise(response, {
-			loading: 'Loading...',
-			success(data) {
-				reset();
-				return 'Đăng ký thành công';
-			},
-			error(error) {
-				reset(
-					{},
-					{
-						keepDirtyValues: true,
-					}
-				);
-				return error;
-			},
-		});
-	}
+	const { register, errors, isSubmitted, handleRegister } = useRegisterForm();
 
 	return (
 		<>
 			<Toaster position="top-center" expand gap={10} closeButton richColors />
-			<form className="w-full space-y-6" onSubmit={handleSubmit(onSubmit)}>
+			<form className="w-full space-y-6" onSubmit={handleRegister}>
 				<Input
 					// disabled={isLoading}
 					register={register}
